requirements.txt
aggforce @ git+https://github.com/noegroup/aggforce.git
ase
e3nn>=0.5.6
h5py
ipython
ipywidgets
jsonargparse[signatures]==4.27.7
matplotlib
mdtraj
mpmath
networkx
nglview
numpy<2.0
packaging
pandas
pytorch_lightning==2.2.1
pyyaml>=6.0
requests
ruamel.yaml
scipy
scikit-learn
sympy
seaborn
torch_optimizer<=1.5.1
tqdm
# developer tools
black[jupyter]
coverage[toml]
pytest
# by mlcg_opt_radius
ninja
# allegro dependencies
nequip==0.12.1
nequip-allegro==0.7.0

